requirements.txt
psutil~=6.1.1
uvicorn~=0.34.0
fastapi~=0.115.6
pyinstaller~=6.11.1
mysql-connector-python~=9.1.0
requests~=2.32.3
